requirements.txt
openai
httpx
python-dotenv
numpy
requests
matplotlib
ipython
weaviate-client
tqdm
joblib
together
flask
scikit-learn
torch
sentence_transformers
ipywidgets
adjusttext
python-dateutil
pandas
dlai-grader
flagembedding
bm25s
phoenix
markdown
opentelemetry-instrumentation-openai
arize-phoenix-otel
beautifulsoup4
deeplake
python-dotenv
faiss-cpu
google-api-python-client
gpt4all
huggingface-hub
langchain
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-huggingface
langchain-openai
langchain-text-splitters
matplotlib
networkx
newspaper3k
nltk
numpy
openai
playwright
pydantic
pyvis
readability-lxml
requests
sentencepiece
scikit-learn
spacy
tiktoken
torch
tqdm
transformers
typing-extensions
openai-whisper
yt-dlp
langchain-classic
langgraph
presidio-anonymizer
presidio-analyzer
datafog
fastapi
uvicorn
fastapi
langgraph-prebuilt
pypdf

vehicle
